Okay, cool, that took all of 5 or 10 seconds longer to sort out.  It had previously been fine to print before the console was set up, and to printf once the heap was set up, and I remember maknig sure the heap was set up, so thinking printf was okay.  But console wasn't set up, so default active console was -1.  So I had silent memory corruption, and all of the strange behavior seems compatible with that.  I think it's likely fine to printTo before console is set up (haven't checked and not sure), but besides removing the offending printf statement I don't need, I'm now also checking that at (active terminal) is set before printing.  (I could default to 0 or 1... but I don't really care right now.)  So now I thing can go back to checking that malloc works the way I was just experimenting with changing to when this all happend.  And then get back to setting up a readline syscall, and then get back to setting up a basic shell in userspace, and get prompting upon process completion working properly, and everthing else I was in the middle of having fun with.
I think there are some broader lessons here around this type of bug and thinking about how to avoid or lessen the likelihood of similar ones in the future.  And around how defeated I felt, and afraid that it was a deeper, fundamental problem with my approach (which, sure, they may be some, but sensing the fear and despair is what's interesting here -- discovering a fundamental problem with my approach would be an opportunity to deepend my understanding, which is a lot of the point of this project), and got ungrounded and spacey.  Getting focused on: okay, so where are we getting stuck when booting fails?  Can't we go slow, do some output, and narrow it down to exactly where it's happening?  And of course, we could, and finally did.